requirements/common.txt
dj-database-url==0.4.1
django-configurations==2.0.0
django-debug-toolbar==1.5.0
django-dotenv==1.4.1
django-extensions==1.7.2
Django==1.10.0
django-polymorphic-tree==1.3
